+++ NEWS.md
@@ -13,6 +13,7 @@ Red Panda C++ Version 0.14.1
   - fix: wrong icons for file associations
   - fix: editor's font size set by ctrl+mouse wheel will be reset by open the option dialog
   - fix: actions not correctly disabled when compile
+  - fix: can't differentiate disabled and enabled buttons, when using contrast icon set
 
 Red Panda C++ Version 0.14.0
   - enhancement: custom icon set ( in the configuration folder)